run.sh: Delay kiosk-browser to let app start first
When using run.sh, I was noticing that kiosk-browser would show an empty white screen and have the following error in the logs (/var/log/syslog):

```
electron: Failed to load URL: http://localhost:3000/ with error: ERR_CONNECTION_REFUSED
```

Delaying kiosk-browser with a sleep seemed to fix the issue, which indicates that the root cause may be that kiosk-browser is trying to load the page before the app frontend server is running. While this isn't a great long-term fix, it may suffice for now.