Change SHV control flow
Rather than having vxsuite call a vxsuite-complete-system script,
bring the core script directly into vxsuite.

This allows us to remove a lot of the previous special-casing for
local dev vs. prod.

In the process, fix a bug that was causing us to display the
UNVERIFIED placeholder hash for locked-down images instead of the
real hash.

Also rename the show-hash-signature vendor function
show-system-hash for consistency with vxsuite.